Placeholder de dropdown con tipo
