+++ src/components/MCQPanel/utils.js
@@ -1,0 +1,5 @@
+export function optionsToBarChartData(options) {
+  return options.map((i) => {
+    return { label: i };
+  });
+}
